Move SlashCommand to the interaction namespace Update readme
